+++ CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.5.0-beta.22](https://github.com/watson-developer-cloud/discovery-components/compare/v1.5.0-beta.21...v1.5.0-beta.22) (2022-02-22)
+
+
+### Features
+
+* add PDFSource support to PDFViewer and DocumentProvider ([#306](https://github.com/watson-developer-cloud/discovery-components/issues/306)) ([b8caee9](https://github.com/watson-developer-cloud/discovery-components/commit/b8caee9e550d75dbc4455fe52e1ab9782c29a0ce))
+
+
+
+
+
+# [1.5.0-beta.21](https://github.com/watson-developer-cloud/discovery-components/compare/v1.5.0-beta.20...v1.5.0-beta.21) (2022-02-22)
+
+
+### Features
+
+* update toolbar style ([#307](https://github.com/watson-developer-cloud/discovery-components/issues/307)) ([62b69e1](https://github.com/watson-developer-cloud/discovery-components/commit/62b69e149b7a316c59c1591f9ae4e00bb517d14a))
+
+
+
+
+
 # [1.5.0-beta.20](https://github.com/watson-developer-cloud/discovery-components/compare/v1.5.0-beta.19...v1.5.0-beta.20) (2022-02-10)